bug fixing

+++ ecom/products/views.py
@@ -7,13 +7,13 @@ from django.core.paginator import Paginator
 from django.views.generic.edit import UpdateView
 from django.urls import reverse_lazy
 from .forms import ProductForm
-
-
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
     
 def home(request):
     products = Product.objects.filter(is_sold = False).order_by('-created_at')
-    paginator = Paginator(products, 4)
+    paginator = Paginator(products, 5)
     page_number = request.GET.get('page', 1)
     productsData = paginator.get_page(page_number)
     
@@ -33,7 +33,7 @@ def get_product(request, slug):
     
     
     
-    other_products = Product.objects.all()
+    other_products = Product.objects.filter(is_sold = False)
     paginator = Paginator(other_products, 5)
     page_number = request.GET.get('p')
     other_products_data = paginator.get_page(page_number)
@@ -75,24 +75,24 @@ def get_all_products(request):
     return render(request, 'products/allproducts.html', context)
 
 
+# @login_required
+# def create_product(request):
+#     if request.method == 'POST':
+#         form = ProductForm(request.POST, request.FILES)
+#         if form.is_valid():
+#             product = form.save(commit=False)
+#             product.owner = request.user  # Set the owner field to the current user
+#             product.save()
+#             # Save product images
+#             for image in request.FILES.getlist('product_images'):
+#                 ProductImages.objects.create(product=product, image=image)
+#             return redirect('home')
+#     else:
+#         form = ProductForm()
+#     return render(request, 'products/create_product.html', {'form': form})
 
-def create_product(request):
-    if request.method == 'POST':
-        form = ProductForm(request.POST, request.FILES)
-        if form.is_valid():
-            product = form.save(commit=False)
-            product.owner = request.user  # Set the owner field to the current user
-            product.save()
-            # Save product images
-            for image in request.FILES.getlist('product_images'):
-                ProductImages.objects.create(product=product, image=image)
-            return redirect('home')
-    else:
-        form = ProductForm()
-    return render(request, 'products/create_product.html', {'form': form})
 
-
-
+@login_required()
 def display_user_product(request):
     if request.user.is_authenticated:
         user_products = Product.objects.filter(owner = request.user).order_by('is_sold')
@@ -100,36 +100,68 @@ def display_user_product(request):
         page_number = request.GET.get('page', 1)
         productsData = paginator.get_page(page_number)
         context = {'products' : productsData}
-        print(user_products)
+        
         return render(request, 'products/myproducts.html', context)
 
+@login_required()
+def create_product(request):
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = ProductForm(request.POST, request.FILES)
+            if form.is_valid():
+                print("user submitted an form")
+                product = form.save(commit=False)
+                product.owner = request.user
+                product.save()
+                
+                    # Save the images
+                images = request.FILES.getlist('product_images')
+                for image in images:
+                    ProductImages.objects.create(product=product, image=image)
+                return redirect('my_product')
+        else:
+            form = ProductForm()
+    
+    return render(request, 'products/create_product.html', {'form': form})
 
 
-    
-    
+@login_required()
 def delete_product(request, uid):
     if request.user.is_authenticated:
         product = get_object_or_404(Product, uid = uid)
-        product.delete()
-    # return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+        if not product.is_sold:
+            product.delete()
+   
     return HttpResponseRedirect(request.META.get('HTTP_REFERER')) 
       
 
 
-
+@login_required
 def update_product(request, slug):
     product = get_object_or_404(Product, slug=slug)
-
+    if not product.owner == request.user:
+        messages.warning(request,"Sorry you cannot edit this product")
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    
     if request.method == 'POST':
+        print(request.FILES)
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
             form.instance.owner = request.user
             form.save()
+            
+            # Save the images
+            images = request.FILES.getlist('product_images')
+            for image in images:
+                
+                ProductImages.objects.create(product=product, image=image)
             return redirect('my_product')
     else:
         form = ProductForm(instance=product)
 
     return render(request, 'products/edit_product.html', {'form': form})
+
+
 
 
     
@@ -138,3 +170,7 @@ def delete_comment(request, uid):
     comment = Comment.objects.get(uid = uid)
     comment.delete()
     return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+
+
+    
+
